third_party/blink/web_tests/flag-specific/highdpi/fast/table/large-shrink-wrapped-width-expected.txt

Content
FAIL:
Expected 1000010 for width, but got 666677. 

<div class="container">
    <div style="display:inline-block; border: 5px solid salmon;" data-expected-width="1000010">
        <table style="width:100%; background-color:green; table-layout:fixed"><tbody><tr><td>Content</td></tr></tbody></table>
    </div>
</div>
	
PASS
	
PASS
	
PASS
